Move LayoutPortlet to layout module (#905)
