+++ uPortal-layout/src/main/java/org/apereo/portal/layout/LayoutPortlet.java
@@ -12,7 +12,7 @@
  * express or implied. See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apereo.portal.rest.layout;
+package org.apereo.portal.layout;
 
 import com.fasterxml.jackson.annotation.JsonRawValue;
 import com.fasterxml.jackson.core.JsonParser;
