+++ uportal-war/src/main/java/org/apereo/portal/rest/LayoutRESTController.java
@@ -23,11 +23,11 @@ import org.apache.commons.logging.LogFactory;
 import org.apereo.portal.IUserPreferencesManager;
 import org.apereo.portal.IUserProfile;
 import org.apereo.portal.layout.IUserLayoutStore;
+import org.apereo.portal.layout.LayoutPortlet;
 import org.apereo.portal.layout.dlm.DistributedUserLayout;
 import org.apereo.portal.portlet.dao.IPortletDefinitionDao;
 import org.apereo.portal.portlet.om.IPortletDefinition;
 import org.apereo.portal.portlet.om.IPortletWindowId;
-import org.apereo.portal.rest.layout.LayoutPortlet;
 import org.apereo.portal.rest.layout.TabListOfNodes;
 import org.apereo.portal.security.IPerson;
 import org.apereo.portal.security.IPersonManager;
